Fixed bug that was failing connections randomly.
It seems I was in some cases making both peers create offers, which would fail because when the offer arrived the peer would be in an invalid state.
I think I copied this code from the signaling example so maybe the bug is there. Most likely, though, is that I messed up in the server refactor. I'm not sure...